Change update baseline logic so that is populated from desired ci run
