+++ tests/microbenchmark/expected/baselines/wh_n150.yaml
@@ -1,6 +1,8 @@
 # tests/microbenchmark/expected/baselines/wh_n150.yaml
-# Calibrated for WH n150 on dedicated runner 'bgd-lab-06' from a single bench invocation.
-# To recalibrate: tests/microbenchmark/tools/calibrate_on_runner.py (must be run on bgd-lab-06).
+# Calibrated for WH n150 from a single bench invocation on dedicated runner 'bgd-lab-06'.
+# To recalibrate: download a benchmark-json-* artifact from a CI run
+# of build-and-run-all-benchmarks.yml and pass its directory to
+# tests/microbenchmark/tools/update_baseline.py via --from-results-dir.
 #
 # Schema (no in-file arch label — the file is dedicated to one arch):
 #   <bench_title>:
@@ -10,11 +12,9 @@ metadata:
   schema_version: 1
   arch: "WH n150"
   runner_hostname: "bgd-lab-06"
-  calibrated_at: "2026-05-20"
-  notes: "Calibrated locally via tests/microbenchmark/tools/calibrate_on_runner.py."
+  calibrated_at: "2026-05-22"
+  notes: "Updated from a CI benchmark JSON artifact via tests/microbenchmark/tools/update_baseline.py."
 
 ClusterConstructor:
-  default: { median_throughput: 247.7, tolerance_pct: 5 }
-  # "from sdesc" intentionally excluded from comparison — too much run-to-run
-  # variance to give a stable signal yet. 
-  # from sdesc: { median_throughput: 219.7, tolerance_pct: 6 }
+  default: { median_throughput: 248.7, tolerance_pct: 5 }
+  from sdesc: { median_throughput: 194.5, tolerance_pct: 5 }
